+++ ruma-appservice-api/src/thirdparty/get_location_for_room_alias/mod.rs
@@ -1,0 +1,3 @@
+//! Endpoint to retrieve an array of third party network locations from a Matrix room alias.
+
+pub mod v1;