+++ ruma-appservice-api/src/thirdparty/get_protocol/mod.rs
@@ -1,0 +1,3 @@
+//! Endpoint to present clients with specific information about the various third party networks that an application service supports.
+
+pub mod v1;