+++ ruma-appservice-api/src/v1.rs
@@ -1,5 +1,0 @@
-//! Endpoints for the r0.x.x versions of the application service API specification
-
-pub mod event;
-pub mod query;
-pub mod thirdparty;